streamlining content set up
